View controller for registering members
Shows a list of enabled users and their token balance.

Tidied up tests

+++ src/main/java/uk/co/vhome/rmj/repositories/PurchaseRepository.java
@@ -10,5 +10,5 @@ import uk.co.vhome.rmj.entities.Purchase;
 public interface PurchaseRepository extends JpaRepository<Purchase, Long>
 {
 	@Query("select sum(p.quantity) from Purchase p where p.username = ?1")
-	Long calculateBalanceForUser(String username);
+	Integer calculateBalanceForUser(String username);
 }
